Explicitly depend on python-openstackclient
currently it is installed only due to some other packages exporting osc
plugins have python-openstackclient as a hard dependency (e.g. ironic)
that should not actually be the case.

Instead, add python-openstackclient as explicit requirement to make sure
we always get it and its core plugins like nova and keystone.

Change-Id: I1215be38e9f47057744303f94f7981d81c37569a