Update openstackclient-check-plugins to honor constraints
Congratulations! You're the first customer of our fine new command pbrx
install-siblings!!!

Srrsly though - the job doesn't currently honor upper-constraints, but
doing that is, it turns out, hard when you're also trying to honor
required-projects. The logic that handles doing that for tox jobs got
extracted into a tool. Use that tool here to install the code from the
required-projects repos - and pass in the constraints file so that
constraints are applied appropriately.

Depends-On: https://review.openstack.org/589476
Change-Id: Ic29d2c1bb15479bd0852b74b5d08aac7ef90048e