Constrain openstackclient-check-plugins run on master
This job is currently running on stable branches, but that is now
resulting in the master version of openstackclient being pulled into the
job. That then will install the latest versions of things, and ends up
failing due to newer libraries being installed than what is in the
stable branch's upper-constraints.

Change-Id: Iedb94a84d2f09c05cc575a955b508e0fd6ca0a64
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>